requirements.txt
Django
djangorestframework
python-dotenv
drf-yasg
psycopg2
django-filter